+++ ESFramework/App/ESApp+Subclassing.m
@@ -20,9 +20,4 @@
     return nil;
 }
 
-- (NSTimeZone *)appWebServerTimeZone
-{
-    return [NSTimeZone timeZoneWithName:@"GMT"];
-}
-
 @end